Put Iam Role Task
Create a new common task to create an iam role with a provided inline policy
that can be assumed by the cluster (nodes-role-arn).

The inline policy should be located in at `source/iam/inline_policy.json`, this
is not configurable at the moment.

To use this common task you need to pass through the following variables:

```bash
AWS_ACCESS_KEY_ID: access key id for aws user that can create/edit policies
AWS_SECRET_ACCESS_KEY: access key for above
AWS_NODES_ROLE_ARN: the role that the nodes that can assume this role will have
APP_NAME: a name for the inline_policy
```

The user needs the following policy attached if you're making it manually; we'll
be creating in terraform.

```json
{
    "Version": "2012-10-17",
    "Statement": [
        {
            "Sid": "VisualEditor0",
            "Effect": "Allow",
            "Action": [
                "iam:GetRole",
                "iam:PutRolePolicy",
                "iam:UpdateAssumeRolePolicy",
                "iam:DeleteRolePolicy",
                "iam:UpdateRoleDescription",
                "iam:CreateRole",
                "iam:AttachRolePolicy",
                "iam:UpdateRole",
                "iam:PutRolePolicy",
                "iam:GetRolePolicy"
            ],
            "Resource": "arn:aws:iam::*:role/*"
        },
        {
            "Sid": "VisualEditor1",
            "Effect": "Allow",
            "Action": "iam:ListRoles",
            "Resource": "*"
        }
    ]
}
```