webapp-docker-image: Experimenting with kaniko resource
> kaniko is a tool to build container images from a Dockerfile [...]
> [...]
> kaniko doesn't depend on a Docker daemon and executes each
> command within a Dockerfile completely in userspace.

Kaniko can cache docker layers by pushing these to a special
docker repository and reusing them later on which is ideal
in the context of Concourse where each task runs in an independent
container.

There is a bug in the kaniko caching mechanism at the moment but there is
a PR which seems to solve it (I tested it locally and it's now using the cached layers).

This task is using a resource which was already there, we'll see if this
works (resource code looks OK to me)

Resourse: https://github.com/lxfontes/kaniko-resource
Kaniko: https://github.com/GoogleContainerTools/kaniko
Kanino cache bugfix: https://github.com/GoogleContainerTools/kaniko/pull/441/