Add support to "port" deploy param
Default to `80` for now to avoid breaking compatibility with existing
webapps.

Once we got to a stage were most apps are updated to add the `port`
param to their `deploy.json` and moved to images that don't run
as `root` we can even start to raise exceptions when this param is
less than `1025`.

Part of ticket: https://trello.com/c/x7zhOJ4N/223-run-webapp-as-non-root-not-on-port-80